+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed vanilla goat horns not playing audio
+- Added web server upload support